% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravClassr.R
\name{gravClassr}
\alias{gravClassr}
\title{This function treats k means clusters as physical objects with mass and gravity,
calculating that gravity for each observation and flagging observations whose
distance and gravity classifications differ.}
\usage{
gravClassr(df, var, k)
}
\arguments{
\item{df}{dataframe}

\item{var}{a column name or list of column names from the df}

\item{k}{number of clusters}
}
\value{

}
\description{
This function treats k means clusters as physical objects with mass and gravity,
calculating that gravity for each observation and flagging observations whose
distance and gravity classifications differ.
}
